requirements.txt
Babel==2.7.0
dj-database-url==0.5.0
Django==2.2.2
django-admin==1.3.2
django-admin-rangefilter==0.5.0
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==0.0.8
django-csp==3.5
django-excel-response2==2.0.8
django-heroku==0.3.1
django-phonenumber-field==3.0.1
django-six==1.0.4
gunicorn==19.9.0
phonenumbers==8.10.15
psycopg2==2.8.3
pytz==2019.2
PyYAML==5.1.2
screen==1.0.1
sqlparse==0.3.0
whitenoise==4.1.3
xlwt==1.3.0
